requirements.txt
pymongo==4.0.1
dnspython==2.1.0
flask
google-cloud-pubsub
confluent-kafka==1.7.0
attrs==21.2.0
avro-python3==1.10.0
jsonschema==3.2.0
requests==2.26.0